+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * [#703](https://github.com/bbatsov/rubocop/issues/703): BracesAroundHashParameters auto-correction broken with trailing comma. ([@jonas054][])
 * [#709](https://github.com/bbatsov/rubocop/issues/709): When `EndAlignment` has configuration `AlignWith: variable`, it now handles `@@a = if ...` and `a, b = if ...`. ([@jonas054][])
 * `SpaceAroundOperators` now reports an offence for `@@a=0`. ([@jonas054][])
+* [#707](https://github.com/bbatsov/rubocop/issues/707): Fix error on operator assignments in top level scope in `UselessAssignment`. ([@yujinakayama][])
 
 ## 0.16.0 (25/12/2013)